Remove unnecessary cast

+++ src/lib.rs
@@ -156,7 +156,7 @@ pub mod utils {
 
     pub fn build_maxpro_lhd(n_samples: usize, n_dim: usize, n_iterations: usize) -> Vec<Vec<f64>> {
         let mut best_metric = f64::INFINITY;
-        let mut best_lhd: Vec<Vec<f64>> = vec![vec![0.0; n_dim]; n_samples as usize];
+        let mut best_lhd: Vec<Vec<f64>> = vec![vec![0.0; n_dim]; n_samples];
         for _i in 0..n_iterations {
             let lhd: Vec<Vec<f64>> = generate_lhd(n_samples, n_dim);
             let maxpro_metric: f64 = maxpro_sum(&lhd);
